+++ FirebaseAI/Sources/Types/Public/Backend.swift
@@ -26,12 +26,7 @@ public struct Backend {
   Use enterprise(location:) instead; note that the default location is now "global" instead of "us-central1"
   """)
   public static func vertexAI() -> Backend {
-    return Backend(
-      apiConfig: APIConfig(
-        service: .enterprise(endpoint: .firebaseProxyProd, location: "us-central1"),
-        version: .v1beta
-      )
-    )
+    return enterprise(location: "us-central1")
   }
 
   /// Initializes a `Backend` configured for the Gemini API in Vertex AI.
@@ -45,17 +40,13 @@ public struct Backend {
   Vertex AI has been renamed to the Gemini Enterprise API.
   """)
   public static func vertexAI(location: String) -> Backend {
-    return Backend(
-      apiConfig: APIConfig(
-        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
-        version: .v1beta
-      )
-    )
+    return enterprise(location: location)
   }
 
   /// Initializes a `Backend` configured for the Gemini Enterprise API.
   ///
-  /// > Note: The Gemini Enterprise API was formerly known as Vertex AI.
+  /// > Note: The Gemini Enterprise API was formerly known as Vertex AI and, briefly,
+  /// > the Agent Platform Gemini API.
   ///
   /// - Parameters:
   ///   - location: The region identifier, defaulting to `global`; see
@@ -71,16 +62,18 @@ public struct Backend {
     )
   }
 
+  /// Initializes a `Backend` configured for the Gemini Enterprise API.
+  ///
+  /// - Parameters:
+  ///   - location: The region identifier, defaulting to `global`; see
+  ///     [available locations][1] for a list of supported locations.
+  ///
+  /// [1]: https://firebase.google.com/docs/ai-logic/locations?api=vertex#available-locations
   @available(*, deprecated, renamed: "enterprise(location:)", message: """
   The Agent Platform Gemini API has been renamed to the Gemini Enterprise API.
   """)
   public static func agentPlatform(location: String = "global") -> Backend {
-    return Backend(
-      apiConfig: APIConfig(
-        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
-        version: .v1beta
-      )
-    )
+    return enterprise(location: location)
   }
 
   /// Initializes a `Backend` configured for the Gemini Developer API.