+++ FirebaseAI/Tests/Unit/TestUtilities/FirebaseAI+DefaultAPIConfig.swift
@@ -15,7 +15,7 @@
 @testable import FirebaseAILogic
 
 extension FirebaseAI {
-  static let defaultVertexAIAPIConfig = APIConfig(
+  static let defaultEnterpriseAPIConfig = APIConfig(
     service: .enterprise(endpoint: .firebaseProxyProd, location: "us-central1"),
     version: .v1beta
   )